more cleanup: - restore pages response models to include 'items' again - rename endpoint to pageSnapshots -> pageUrlCounts - fix tests to not check 'total'
